backend/pyproject.toml
[tool.poetry]
name = "university"
version = "0.1.0"
description = ""
authors = ["Kirill <fikfack806@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic-settings = "^2.6.1"
greenlet = "^3.1.1"
alembic = "^1.14.0"
tenacity = "^9.0.0"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
